rename: StudyService 위치 변경
