+++ src/test/java/com/grepp/spring/app/model/study/StudyServiceTest.java
@@ -4,6 +4,7 @@ package com.grepp.spring.app.model.study;
 import com.grepp.spring.app.model.member.dto.response.ApplicantsResponse;
 import com.grepp.spring.app.model.study.code.ApplicantState;
 import com.grepp.spring.app.model.study.repository.StudyRepository;
+import com.grepp.spring.app.model.study.service.StudyService;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
